Update ai.ts
Updated the OpenAI's response handling to pick the right handler based on the `streaming` flag. This is done through an interface function which has 2 separate implementations: 1 for handling streaming responses and 1 for handling non-streaming responses. And the `responseHandler` which is in charge of assigning the appropriate function implmenetation based on the value of the `streaming` flag.