Update ai.ts (#304)
* Update ai.ts

Updated the OpenAI's response handling to pick the right handler based on the `streaming` flag. This is done through an interface function which has 2 separate implementations: 1 for handling streaming responses and 1 for handling non-streaming responses. And the `responseHandler` which is in charge of assigning the appropriate function implmenetation based on the value of the `streaming` flag.

* Update ai.ts

Fixed lint errors

* Made the changes based on the feedback

Changed responseHandler to handleResponse and removed the return type. Ran pnpm prettier. Fixed the order of type imports. Removed the interface to narrow the typing. Removed the return type from handleStreamingResponse and handleNonStreamingResponse. Simplified to just use the handleResponse function name in .then(). simplified to just use the handleError function name in .catch().