deno routing to try to make existing cloudflare code run (#773)
This is a bunch of hacks to make existing cloudflare code run in Deno.

Main problem is file-system routing, we'll need to convert this to static routing some day. Also need to fix imports some day. For now this can run with smallweb/deno locally, but not on deno deploy due to needing sloppy imports and "dynamic" fs routing.